Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -7,31 +7,31 @@
 int main() {
      
     
-    int num1,num2,result;
+    float num1,num2,result;
     char calculator;
-    scanf("%d %d",&num1,&num2);
+    scanf("%f %f",&num1,&num2);
     scanf("%c",&calculator);
 
     switch(calculator){
         case '+':
         result = num1 + num2;
-        printf("%d\n",result);
+        printf("%f\n",result);
         break;
 
         case '-':
         result=num1-num2;
-        printf("%d\n",result);
+        printf("%f\n",result);
         break;
 
         case '*':
         result=num1*num2;
-        printf("%d\n",result);
+        printf("%f\n",result);
         break;
 
         case '/':
          if(num2 != 0){
             result= num1/num2;
-            printf("%d\n",result);
+            printf("%f\n",result);
             }
             else{
                 printf("Error");
